:bug: :gun: Help template figure out value

+++ 2022/Rust/.template/src/part_one/mod.rs
@@ -7,7 +7,7 @@ pub fn run(a: args::Arguments) {
     );
 }
 
-fn run_string(s: String) -> u32 {
+fn run_string(s: String) -> usize {
     println!("{}", s);
     0
 }
